Add support for aix and solaris.
